+++ src/pytickersymbols/__init__.py
@@ -10,7 +10,7 @@ import json
 import yaml
 from weakref import WeakValueDictionary
 
-__version__ = "1.3.0"
+__version__ = "1.7.8"
 
 
 class Singleton(type):
@@ -32,7 +32,7 @@ class PyTickerSymbols(metaclass=Singleton):
                 'data',
                 'stocks.json',
             )
-            with open(json_path) as stocks:
+            with open(json_path, errors='replace') as stocks:
                 self.__stocks = json.load(stocks)
         else:
             if stocks_path.lower().endswith(
@@ -67,6 +67,59 @@ class PyTickerSymbols(metaclass=Singleton):
         :return: list of index names
         """
         return self.__get_sub_items('indices')
+
+    def get_all_stocks(self):
+        """
+        Returns all available stocks
+        :return: list of index stock objects
+        """
+        return self.__stocks['companies']
+
+    def get_stock_name_by_yahoo_symbol(self, symbol):
+        """
+        Returns stock name by yahoo symbol
+        :return: stock name or None if symbol is not present
+        """
+        stock = self.get_stock_by_yahoo_symbol(symbol)
+        return stock['name'] if stock else None
+
+    def get_stock_name_by_google_symbol(self, symbol):
+        """
+        Returns stock name by google symbol
+        :return: stock name or None if symbol is not present
+        """
+        stock = self.get_stock_by_google_symbol(symbol)
+        return stock['name'] if stock else None
+
+    def get_stock_by_yahoo_symbol(self, symbol):
+        """
+        Returns stock by yahoo symbol
+        :return: stock or None if symbol is not present
+        """
+        return self.__get_stock_by_symbol(symbol, 'yahoo')
+
+    def get_stock_by_google_symbol(self, symbol):
+        """
+        Returns stock by google symbol
+        :return: stock or None if symbol is not present
+        """
+        return self.__get_stock_by_symbol(symbol, 'google')
+
+    def __get_stock_by_symbol(self, symbol, symbol_type):
+        """
+        Returns stock name by symbol
+        :return: stock name or None if symbol is not present
+        """
+        return next(
+            filter(
+                lambda x, sym=symbol: sym
+                in map(
+                    lambda y, sym_type=symbol_type: y[sym_type], x['symbols']
+                ),
+                self.get_all_stocks(),
+            ),
+            None
+        )
 
     def get_all_industries(self):
         """