+++ tests/test_lib.py
@@ -14,7 +14,8 @@ import tempfile
 import json
 import yaml
 import pytest
-
+from collections import Counter
+import pycountry
 
 _test_json = b'''{
    "indices": [
@@ -52,7 +53,6 @@ _test_json = b'''{
 
 
 class TestLib(unittest.TestCase):
-
     def test_singleton(self):
         """
         Test singleton pattern
@@ -82,7 +82,11 @@ class TestLib(unittest.TestCase):
     def test_load_yaml(self):
         for sufix_test in ['.yaml', '.yml']:
             with tempfile.NamedTemporaryFile(suffix=sufix_test) as temp:
-                temp.write(str.encode(yaml.dump(json.loads(_test_json.decode('utf-8')))))
+                temp.write(
+                    str.encode(
+                        yaml.dump(json.loads(_test_json.decode('utf-8')))
+                    )
+                )
                 temp.flush()
                 stocks = PyTickerSymbols(stocks_path=temp.name)
                 indices = stocks.get_all_indices()
@@ -102,13 +106,23 @@ class TestLib(unittest.TestCase):
         self.assertIsNotNone(stock_data)
         indices = stock_data.get_all_indices()
         self.assertIsNotNone(indices)
-        self.assertIn("DAX", indices)
-        self.assertIn("SDAX", indices)
-        self.assertIn("MDAX", indices)
+        self.assertIn('DAX', indices)
+        self.assertIn('SDAX', indices)
+        self.assertIn('MDAX', indices)
         # duplicates are not allowed
         for index in indices:
             lenl = len([tmp for tmp in indices if tmp == index])
             self.assertEqual(lenl, 1)
+
+    def test_all_stocks(self):
+        """
+        Test stocks getter
+        :return:
+        """
+        stock_data = PyTickerSymbols()
+        self.assertIsNotNone(stock_data)
+        stocks = stock_data.get_all_stocks()
+        self.assertTrue(len(stocks) > 900)
 
     def test_encoding(self):
         """
@@ -117,8 +131,19 @@ class TestLib(unittest.TestCase):
         """
         stock_data = PyTickerSymbols()
         self.assertIsNotNone(stock_data)
-        dax = list(stock_data.get_stocks_by_index('DAX'))
-        self.assertEqual(dax[10]['name'], 'Deutsche Börse AG')
+        stocks = list(stock_data.get_all_stocks())
+        names_with_unicode_count = sum(
+            [
+                any(
+                    word in item['name']
+                    for word in ['ö', 'ä', 'ü', 'é', 'è', 'ë']
+                )
+                for item in stocks
+            ]
+        )
+        self.assertTrue(
+            names_with_unicode_count >= 20, names_with_unicode_count
+        )
 
     def test_country(self):
         """
@@ -181,8 +206,12 @@ class TestLib(unittest.TestCase):
                 self.assertTrue(is_in)
         # test NASDAQ 100
         stocks_nasdaq = list(stock_data.get_stocks_by_index('NASDAQ 100'))
-        stocks_nasdaq_symbol = [sym['yahoo'] for stock in stocks_nasdaq for sym in stock['symbols']]
-        symbols_nasdaq = list(stock_data.get_yahoo_ticker_symbols_by_index('NASDAQ 100'))
+        stocks_nasdaq_symbol = [
+            sym['yahoo'] for stock in stocks_nasdaq for sym in stock['symbols']
+        ]
+        symbols_nasdaq = list(
+            stock_data.get_yahoo_ticker_symbols_by_index('NASDAQ 100')
+        )
         symbols_nasdaq = reduce(lambda x, y: x + y, symbols_nasdaq)
         self.assertEqual(len(stocks_nasdaq_symbol), len(symbols_nasdaq))
         self.assertIn('GOOGL', symbols_nasdaq)
@@ -242,6 +271,58 @@ class TestLib(unittest.TestCase):
                     is_in_basic = True
             self.assertTrue(is_in_basic)
 
+    def test_stock_by_yahoo_symbol(self):
+        """
+        Tests stock getter by industry
+        :return:
+        """
+        stock_data = PyTickerSymbols()
+        self.assertIsNotNone(stock_data)
+        ads = stock_data.get_stock_by_yahoo_symbol('ADS.F')
+        self.assertIsNotNone(ads)
+        self.assertEqual('adidas AG', ads['name'])
+        unknown = stock_data.get_stock_by_yahoo_symbol('ADSdadsas.F')
+        self.assertIsNone(unknown)
+
+    def test_stock_name_by_yahoo_symbol(self):
+        """
+        Tests stock getter by industry
+        :return:
+        """
+        stock_data = PyTickerSymbols()
+        self.assertIsNotNone(stock_data)
+        ads = stock_data.get_stock_name_by_yahoo_symbol('ADS.F')
+        self.assertIsNotNone(ads)
+        self.assertEqual('adidas AG', ads)
+        unknown = stock_data.get_stock_by_yahoo_symbol('ADSdadsas.F')
+        self.assertIsNone(unknown)
+
+    def test_stock_by_google_symbol(self):
+        """
+        Tests stock getter by industry
+        :return:
+        """
+        stock_data = PyTickerSymbols()
+        self.assertIsNotNone(stock_data)
+        ads = stock_data.get_stock_by_google_symbol('FRA:ADS')
+        self.assertIsNotNone(ads)
+        self.assertEqual('adidas AG', ads['name'])
+        unknown = stock_data.get_stock_by_google_symbol('ADSdadsas.F')
+        self.assertIsNone(unknown)
+
+    def test_stock_name_by_google_symbol(self):
+        """
+        Tests stock getter by industry
+        :return:
+        """
+        stock_data = PyTickerSymbols()
+        self.assertIsNotNone(stock_data)
+        ads = stock_data.get_stock_name_by_google_symbol('FRA:ADS')
+        self.assertIsNotNone(ads)
+        self.assertEqual('adidas AG', ads)
+        unknown = stock_data.get_stock_by_google_symbol('ADSdadsas.F')
+        self.assertIsNone(unknown)
+
     def test_tickers_by_index(self):
         """
         Tests tickers getter by index
@@ -273,7 +354,7 @@ class TestLib(unittest.TestCase):
         for test_item in test_list:
             self.assertEqual(len(test_item), 30)
             for tickers in test_item:
-                self.assertEqual(len(tickers), 2)
+                self.assertTrue(len(tickers) >= 1, tickers)
 
     def test_tickers_valid(self):
         """
@@ -297,6 +378,67 @@ class TestLib(unittest.TestCase):
         self.assertEqual('^MDAXI', stock_data.index_to_yahoo_symbol('MDAX'))
         swi = stock_data.index_to_yahoo_symbol('Switzerland 20')
         self.assertEqual('^SSMI', swi)
+
+    def test_unique_ticker_symbols(self):
+        stock_data = PyTickerSymbols()
+        ctx = Counter(
+            [
+                sym['yahoo']
+                for stock in stock_data.get_all_stocks()
+                for sym in stock['symbols']
+            ]
+        )
+        msg = 'The following symbols appear several times:\n'
+        msg += '\n'.join(
+            map(lambda y: y[0], filter(lambda x: x[1] > 1, ctx.items()))
+        )
+        self.assertFalse(any(map(lambda x: x > 1, ctx.values())), msg)
+
+    def test_valid_country_name(self):
+        stock_data = PyTickerSymbols()
+        countries = stock_data.get_all_countries()
+        empty_names = list(filter(lambda x: not x, countries))
+        empty_country_stocks = ', '.join(
+            list(
+                map(
+                    lambda x: x['name'],
+                    stock_data.get_stocks_by_country(''),
+                )
+            )
+        )
+        self.assertEqual(
+            len(empty_names),
+            0,
+            'The following stocks have an empty country string: '
+            + empty_country_stocks,
+        )
+
+        valid_countires = list(
+            map(
+                lambda x: x.name,
+                pycountry.countries,
+            )
+        )
+        wrong_country_name = list(
+            filter(
+                lambda x: x['country'] not in valid_countires,
+                stock_data.get_all_stocks(),
+            ),
+        )
+        wrong_country_name_stocks = ', '.join(
+            list(
+                map(
+                    lambda x: x['name'] + '(' + x['country'] + ')',
+                    wrong_country_name,
+                )
+            )
+        )
+        self.assertEqual(
+            len(wrong_country_name),
+            0,
+            'The following stocks have an empty country string:'
+            + wrong_country_name_stocks,
+        )
 
 
 if __name__ == "__main__":