+++ CHANGELOG.md
@@ -2,6 +2,66 @@
 
 <!--next-version-placeholder-->
 
+## v1.7.8 (2021-07-21)
+### Fix
+* Update stock symbols ([`abcc7d0`](https://github.com/portfolioplus/pytickersymbols/commit/abcc7d01f81b85f1e3428924999d21ca903ef2c0))
+
+## v1.7.7 (2021-07-13)
+### Fix
+* Update stock symbols ([`c8dd41c`](https://github.com/portfolioplus/pytickersymbols/commit/c8dd41ca9a462291d7a75dfb362966b26a3c0a40))
+
+## v1.7.6 (2021-06-16)
+### Fix
+* Update stock symbols ([`f295a48`](https://github.com/portfolioplus/pytickersymbols/commit/f295a487360b95776aeedc1cfa2df0f074915195))
+
+## v1.7.5 (2021-06-09)
+### Fix
+* Update stock symbols ([`2ca408a`](https://github.com/portfolioplus/pytickersymbols/commit/2ca408a973eed863e24800158874d48face2356d))
+
+## v1.7.4 (2021-06-09)
+### Fix
+* Removed anchors ([`3fd1af1`](https://github.com/portfolioplus/pytickersymbols/commit/3fd1af1e35a2db73a920ca4e5d77755161719abc))
+
+## v1.7.3 (2021-06-09)
+### Fix
+* Removed wrong metadata ([`3c8f46e`](https://github.com/portfolioplus/pytickersymbols/commit/3c8f46ec3d239ee256b080c77e0444b22a7b935b))
+
+## v1.7.2 (2021-06-09)
+### Fix
+* Added wiki name for wdp ([`ab39dbf`](https://github.com/portfolioplus/pytickersymbols/commit/ab39dbfdd2af4a6c92126410c1765669e67aa562))
+
+## v1.7.1 (2021-06-08)
+### Fix
+* Removed duplicate IMCD ([`9083ab5`](https://github.com/portfolioplus/pytickersymbols/commit/9083ab5c47016227482eb829403565f2d1598d40))
+
+## v1.7.0 (2021-06-04)
+### Feature
+* Updated stocks and removed ids ([`355a998`](https://github.com/portfolioplus/pytickersymbols/commit/355a998bc50c90427f7c6c386af941acb8d86eb4))
+
+## v1.6.0 (2021-06-03)
+### Feature
+* Unicode characters added to yaml file to allow human reading ([`6dd38a7`](https://github.com/portfolioplus/pytickersymbols/commit/6dd38a7cc1aae245e015859077355b2900ef8d46))
+
+## v1.5.0 (2021-05-26)
+### Feature
+* Get stock by ticker symbol ([#85](https://github.com/portfolioplus/pytickersymbols/issues/85)) ([`47662e8`](https://github.com/portfolioplus/pytickersymbols/commit/47662e85deeb71f27cb8596be1838dd3ae0fdb6b))
+
+## v1.4.3 (2021-01-19)
+### Fix
+* Country names ([`8ff8538`](https://github.com/portfolioplus/pytickersymbols/commit/8ff8538e630055b039ee84382fc9476cec201bfc))
+
+## v1.4.2 (2021-01-11)
+### Fix
+* Removed duplicate symbols ([`8bfd119`](https://github.com/portfolioplus/pytickersymbols/commit/8bfd1199871a0757a0820ab948250261fab1a81d))
+
+## v1.4.1 (2021-01-08)
+### Fix
+* Update stock symbols ([`2667419`](https://github.com/portfolioplus/pytickersymbols/commit/266741998b114477c3587705797e5d833ecc9564))
+
+## v1.4.0 (2020-12-16)
+### Feature
+* Get all stocks getter ([#58](https://github.com/portfolioplus/pytickersymbols/pull/58)) ([`2c171ba`](https://github.com/portfolioplus/pytickersymbols/commit/2c171ba0a04d39ba0d19bf9bb06aab0d5fc10471))
+
 ## v1.3.0 (2020-11-26)
 ### Feature
 * Load external (#51) ([`8c9cdb1`](https://github.com/portfolioplus/pytickersymbols/commit/8c9cdb1e755f79206876c6138e508452a29f46f5))